+++ assets/src/dashboard/utils/index.js
@@ -15,7 +15,6 @@
  */
 export { default as titleFormatted } from './titleFormatted';
 export { default as keyboardOnlyOutline } from './keyboardOnlyOutline';
-export { default as memoize } from './memoize';
 export { default as throttleToAnimationFrame } from './throttleToAnimationFrame';
 export { default as useDashboardResultsLabel } from './useDashboardResultsLabel';
 export { default as useFocusOut } from './useFocusOut';
