+++ package.json
@@ -41,6 +41,7 @@
     "@web-stories-wp/jest-amp": "*",
     "@web-stories-wp/jest-puppeteer-amp": "*",
     "@web-stories-wp/migration": "*",
+    "@web-stories-wp/templates": "*",
     "@web-stories-wp/tracking": "*",
     "@wordpress/api-fetch": "^3.21.3",
     "@wordpress/block-editor": "^5.2.1",
@@ -259,8 +260,8 @@
     "workflow:migrate": "npm-run-all --parallel workflow:migrate:*",
     "workflow:migrate:text-sets": "migration assets/src/edit-story/components/library/panes/text/textSets/raw",
     "postworkflow:migrate:text-sets": "npx prettier --write assets/src/edit-story/components/library/panes/text/textSets/raw",
-    "workflow:migrate:templates": "migration assets/src/dashboard/templates/raw",
-    "postworkflow:migrate:templates": "npx prettier --write assets/src/dashboard/templates/raw",
+    "workflow:migrate:templates": "migration packages/templates/src/raw",
+    "postworkflow:migrate:templates": "npx prettier --write packages/templates/src/raw",
     "workflow:migrate:ftue-story": "migration includes/data/stories",
     "postworkflow:migrate:ftue-story": "npx prettier --write includes/data/stories"
   },
