+++ packages/templates/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@web-stories-wp/templates",
+  "private": true,
+  "engines": {
+    "node": ">= 15",
+    "npm": "~7"
+  },
+  "exports": {
+    "default": "./src/index.js"
+  },
+  "main": "./src/index.js",
+  "type": "module",
+  "dependencies": {
+    "@web-stories-wp/migration": "*",
+    "@web-stories-wp/tracking": "*",
+    "@web-stories-wp/i18n": "*"
+  },
+  "devDependencies": {}
+}
