+++ packages/templates/src/raw/beauty.json
