+++ packages/templates/src/raw/cooking.json
