+++ packages/templates/src/raw/diy.json
