+++ packages/templates/src/raw/entertainment.json
