+++ packages/templates/src/raw/fashion.json
