+++ packages/templates/src/raw/fitness.json
