+++ packages/templates/src/raw/travel.json
