+++ packages/templates/src/raw/wellbeing.json
