+++ packages/templates/src/utils/memoize.js
